MetadataEntity redesign

+++ app/components/impl/src/main/java/com/example/healthconnect/components/impl/ui/metadata/MetadataEditorViewModel.kt
@@ -29,8 +29,7 @@ class MetadataEditorViewModel(
             }
 
             is Event.OnClientVersionChanged -> {
-                //TODO check input
-                _state.copy(clientRecordVersion = event.value.toLong())
+                _state.copy(clientRecordVersion = event.value)
             }
 
             Event.OnSpecifyDevice -> _state.copy(
